Refactor generic insertion and query
